backend/requirements.txt
django>=4.2.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.1.0
django-filter>=23.2
drf-spectacular>=0.26.3
django-extensions>=3.2.3
